OneDrive/Desktop/ProjectWk/FITT/Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.28.0
python-dotenv==1.0.1
pydantic[email]
requests==2.32.3
groq==1.0.0
python-multipart>=0.0.7